Move ExceptionMapper(s) to a separate namespace
