+++ src/Opw.HttpExceptions.AspNetCore/Mappers/IExceptionMapper.cs
@@ -2,7 +2,7 @@ using System;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Opw.HttpExceptions.AspNetCore
+namespace Opw.HttpExceptions.AspNetCore.Mappers
 {
     /// <summary>
     /// Interface for ExceptionMappers.
